+++ R/conos.R
@@ -1,6 +1,7 @@
 #' @useDynLib conos
 #' @import Matrix
 #' @import igraph
+#' @import sccore
 #' @importFrom parallel mclapply
 #' @importFrom dplyr %>%
 #' @importFrom magrittr %<>%
@@ -281,7 +282,7 @@ quickPlainPCA <- function(r.n,data.type='counts',ncomps=30,n.odgenes=NULL,var.sc
 }
 
 
-# Perform CCA (using PMA package or otherwise) on two samples
+#' Perform CCA (using PMA package or otherwise) on two samples
 #' @param r.n list of p2 objects
 #' @param ncomps number of components to calculate (default=100)
 #' @param n.odgenes number of overdispersed genes to take from each dataset
@@ -990,67 +991,6 @@ reduceEdgesInGraphIteratively <- function(adj.mtx,k,preserve.order=TRUE,max.kdif
   }
 
   return(adj.mtx)
-}
-
-##' Collapse vertices belonging to each cluster in a graph
-##'
-##' @param graph graph to be collapsed
-##' @param groups factor on vertives describing cluster assignment (can specify integer vertex ids, or character vertex names which will be matched)
-##' @param plot whether to show collapsed graph plot
-##' @param normalize whether recalculate edge weight as observed/oexpected
-##' @return collapsed graph
-getClusterGraph <- function(graph,groups,plot=FALSE,node.scale=50,edge.scale=50,edge.alpha=0.3,normalize=TRUE) {
-  V(graph)$num <- 1;
-  if(is.integer(groups) && is.null(names(groups))) {
-    nv <- vcount(graph)
-    if(length(groups)!=nv) stop('length of groups should be equal to the number of vertices')
-    if(max(groups)>nv) stop('groups specifies ids that are larger than the number of vertices in the graph')
-    if(any(is.na(groups))) {
-      # remove vertices that are not part of the groups
-      vi <- which(!is.na(groups));
-      g <- induced.subgraph(graph,vi);
-      groups <- groups[vi];
-    } else {
-      g <- graph;
-    }
-  } else {
-    gn <- V(graph)$name;
-    groups <- na.omit(groups[names(groups) %in% gn]);
-    if(length(groups)<2) stop('valid names of groups elements include too few cells')
-    if(length(groups)<length(gn)) {
-      g <- induced.subgraph(graph,names(groups))
-    } else {
-      g <- graph;
-    }
-    if(is.factor(groups)) {
-      groups <- groups[V(g)$name]
-    } else {
-      groups <- as.factor(setNames(as.character(groups[V(g)$name]),V(g)$name))
-    }
-  }
-  gcon <- contract.vertices(g,groups,vertex.attr.comb=list('num'='sum',"ignore"))
-  # translate into observed/expected
-  gcon <- simplify(gcon, edge.attr.comb=list(weight="sum","ignore"))
-  if(normalize) {
-    ex <- outer(V(gcon)$num,V(gcon)$num)/(sum(V(gcon)$num)*(sum(V(gcon)$num)-1)/2)*sum(E(g)$weight)
-    gcon2 <- graph_from_adjacency_matrix(as(as_adjacency_matrix(gcon,attr = "weight",sparse = F)/ex,'dgCMatrix'),mode = "undirected",weighted=TRUE)
-    V(gcon2)$num <- V(gcon)$num
-    gcon <- gcon2;
-  }
-  if(is.factor(groups)) {
-    V(gcon)$name <- levels(groups)
-  } else {
-    # not sure when this was actually needed
-    gcon <- induced.subgraph(gcon, unique(groups))
-  }
-
-
-  if(plot) {
-    set.seed(1)
-    par(mar = rep(0.1, 4))
-    plot.igraph(gcon, layout=layout_with_fr(gcon), vertex.size=V(gcon)$num/(sum(V(gcon)$num)/node.scale), edge.width=E(gcon)$weight/sum(E(gcon)$weight/edge.scale), edge.color=adjustcolor('black',alpha=edge.alpha))
-  }
-  return(invisible(gcon))
 }
 
 adjustWeightsByCellBalancing <- function(adj.mtx, factor.per.cell, balance.weights, same.factor.downweight=1.0, n.iters=50, verbose=F) {